Remove empty object creation
rbbi needs to be created with "self", so we can't just create it in the
initializer. Instead, make it a forced unwrapped optional (so it
defaults to nil instead of an empty bar button item but we don't have to
bother using it as a "real" optional since we know it will always have a
value).